Update manifest command to use flux-local
