+++ requirements.txt
@@ -2,6 +2,8 @@
 GitPython==3.1.30
 aiofiles==22.1.0
 iniconfig==2.0.0
+flux-local==0.0.2
+pip==23.0
 platformdirs==2.6.2
 pre-commit==3.0.4
 py==1.11.0
